+++ tests/test_cram_archiver.py
@@ -141,7 +141,7 @@ def test_find_bam_files(tmp_path, caplog, debug):
     os.utime(bam1, (1000, 300))
     os.utime(bam2, (1000, 200))
     os.utime(bam3, (1000, 100))
-    result = list(find_bam_files(str(tmp_path), older_than_timestamp=201))
+    result = list(find_bam_files([str(tmp_path)], older_than_timestamp=201))
     assert set(result) == {str(bam2), str(bam3)}
     assert str(bam1) in caplog.text
     assert (str(bam2) in caplog.text) is debug
@@ -175,7 +175,9 @@ def test_find_bam_files_no_symlinks(tmp_path, caplog, debug):
     os.utime(bam1, (1000, 300))
     os.utime(bam2, (1000, 200))
     os.utime(bam3, (1000, 100))
-    result = list(find_bam_files(str(tmp_path), older_than_timestamp=201))
+    result = list(find_bam_files(
+        [str(bam1_link), str(tmp_path)],
+        older_than_timestamp=201))
     assert set(result) == {str(bam2), str(bam3)}
     assert str(bam1) in caplog.text
     assert (str(bam2) in caplog.text) is debug
@@ -203,7 +205,7 @@ def test_find_bam_files_exclude(tmp_path: Path, caplog, debug):
     decoy1.touch()
     decoy2.touch()
     result = list(find_bam_files(
-        str(tmp_path),
+        [str(tmp_path)],
         # Make the timestamp very big to avoid testing issues.
         older_than_timestamp=math.inf,
         ignore_files=[str(bam1.absolute())]
@@ -236,7 +238,7 @@ def test_find_bam_files_exclude_ext(tmp_path: Path, caplog, debug):
     decoy1.touch()
     decoy2.touch()
     result = list(find_bam_files(
-        str(tmp_path),
+        [str(tmp_path)],
         # Make the timestamp very big to avoid testing issues.
         older_than_timestamp=math.inf,
         ignore_extensions=[".repeats.bam"]
@@ -248,6 +250,47 @@ def test_find_bam_files_exclude_ext(tmp_path: Path, caplog, debug):
     assert (str(bam3) in caplog.text) is debug
     assert (str(decoy1) in caplog.text) is debug
     assert (str(decoy2) in caplog.text) is debug
+
+
+def test_find_bam_files_multiple_paths(tmp_path: Path, caplog):
+    caplog.set_level(logging.INFO)
+    subdir = tmp_path / "subdir"
+    subdir.mkdir()
+    bam1 = tmp_path / "bam1.bam"
+    bam2 = tmp_path / "bam2.bam"
+    bam3 = subdir / "bam3.bam"
+    decoy1 = subdir / "decoy1.txt"
+    decoy2 = tmp_path / "decoy2.txt"
+    bam1.touch()
+    bam2.touch()
+    bam3.touch()
+    decoy1.touch()
+    decoy2.touch()
+    result = list(find_bam_files(
+        [str(bam1), str(tmp_path), str(bam3)],
+        # Make the timestamp very big to avoid testing issues.
+        older_than_timestamp=math.inf,
+        ignore_files=[str(bam1)]
+    ))
+    result.sort()
+    assert result == [str(bam2), str(bam3)]
+    assert str(bam1) in caplog.text
+    assert "Ignoring" in caplog.text
+    assert "WARNING" in caplog.text
+    assert f"Skipping duplicate path: {str(bam3)}" in caplog.text
+
+
+def test_find_bam_files_crash_on_string(tmp_path):
+    bam1 = tmp_path / "bam1.bam"
+    bam1.touch()
+    with pytest.raises(TypeError):
+        list(find_bam_files(str(tmp_path)))  # type: ignore
+
+
+def test_find_bam_files_skip_non_file(tmp_path):
+    fifo_bam = tmp_path / "fifo.bam"
+    os.mkfifo(fifo_bam)
+    assert [] == list(find_bam_files([str(fifo_bam)]))
 
 
 @pytest.mark.parametrize(
@@ -331,7 +374,7 @@ def test_cram_archiver(
         cram_archiver_main(*args)
     else:
         cram_archiver(
-            input_path=str(tmp_path),
+            input_paths=[str(tmp_path)],
             reference_files=[str(TEST_DATA / "NC012920.1.fasta")],
             cram_version=cram_version,
             write_index=write_index,
@@ -423,7 +466,7 @@ def test_cram_archiver_single_file(
         cram_archiver_main(*args)
     else:
         cram_archiver(
-            input_path=str(bam),
+            input_paths=[str(bam)],
             reference_files=[str(TEST_DATA / "NC012920.1.fasta")],
             cram_version=cram_version,
             write_index=write_index,
@@ -461,7 +504,7 @@ def test_cram_archiver_dry_run(tmp_path, capsys, delete,
     os.utime(bam2, (current_time, current_time - 100_000))  # More than 1 day
     os.utime(bam3, (current_time, current_time - 200_000))  # More than 2 days
     cram_archiver(
-        input_path=str(tmp_path),
+        input_paths=[str(tmp_path)],
         reference_files=[str(TEST_DATA / "NC012920.1.fasta")],
         minimum_age_days=1,
         dry_run=True,
@@ -486,7 +529,7 @@ def test_cram_archiver_dry_run_exclude_files(tmp_path, capsys):
     bam2.touch()
     bam3.touch()
     cram_archiver(
-        input_path=str(tmp_path),
+        input_paths=[str(tmp_path)],
         reference_files=[str(TEST_DATA / "NC012920.1.fasta")],
         dry_run=True,
         ignore_files=[str(bam1), str(bam3)],
@@ -508,7 +551,7 @@ def test_cram_archiver_dry_run_root_dir_excluded(tmp_path, capsys):
     bam2.touch()
     bam3.touch()
     cram_archiver(
-        input_path=str(tmp_path),
+        input_paths=[str(tmp_path)],
         reference_files=[str(TEST_DATA / "NC012920.1.fasta")],
         dry_run=True,
         ignore_files=[str(tmp_path)],
@@ -525,7 +568,7 @@ def test_cram_archiver_dry_run_root_file_excluded(tmp_path, capsys):
     bam1 = tmp_path / "bam1.bam"
     bam1.touch()
     cram_archiver(
-        input_path=str(bam1),
+        input_paths=[str(bam1)],
         reference_files=[str(TEST_DATA / "NC012920.1.fasta")],
         dry_run=True,
         ignore_files=[str(bam1)],
@@ -595,7 +638,7 @@ def test_cram_archiver_no_reference_fai(tmp_path):
 
 
 def test_cram_archiver_no_bam_files(tmp_path, caplog):
-    cram_archiver(str(tmp_path), [str(TEST_DATA / "NC012920.1.fasta")])
+    cram_archiver([str(tmp_path)], [str(TEST_DATA / "NC012920.1.fasta")])
     assert "No BAM files found." in caplog.text
 
 
@@ -645,7 +688,7 @@ def test_cram_archiver_fails_gracefully(
     os.utime(bam3, (current_time, current_time - 200_000))  # More than 2 days
     with pytest.raises(RuntimeError) as error:
         cram_archiver(
-            input_path=str(tmp_path),
+            input_paths=[str(tmp_path)],
             reference_files=[str(TEST_DATA / "NC012920.1.fasta")],
             cram_version=cram_version,
             write_index=True,